RoiEdit debug New common abstract file for roi Edit Small file organisation New common abstract file for roi
